frontend/requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
openai
PyYAML
gitpython
markdown
tiktoken